storage: model type relationships through traits
Add a trait that encodes the 1-1 relationship between a
`SourceConnection` and its corresponding `SourceReader` implementation.

Defining it this way instead of the other way around as it used to be
requires less manual type annotations as there is only one reader that
can be instantiated for a particular source connection.

This PR also fixes some generic parameters that were under-constrained so
that type inference can figure them out instead.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>