storage: punctuation response for strict id checks
This PR adds a new type of storage response that serves as punctuation
to let consumers of the response stream know that there will be no
further updates for a particular id.

This is the moment where clients in the client stack can clean up their
state and forget about the existence of an id.

Since dropping an id is now an explicit response clients can be strict
about how the protocol behaves and panic if we ever deviate which would
indicate a bug.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>