Clean up of frontier introspection events during reconciliation (#15967)
It was observed that executing reconciliation after the crash of environmentd with existing computed's, the introspection source mz_worker_compute_frontiers was left in an inconsistent state where excessive retractions had been issued with a logical timestamp of zero.

This commit resolves this problem by fixes to the replica-side reconciliation code. In particular, the reconciliation code was resetting replica-side metadata in ComputeState regarding reported_frontiers, but did not consider that this reset had to be aligned with reporting to the corresponding logging source.