Regression testing for invalid retractions from `mz_worker_compute_frontiers` upon reconciliation (#15986)
We observed that invalid retractions for the introspection source mz_worker_compute_frontiers could be triggered by reconciliation. This commit provides the missing regression test for this issue in the form of a mzcompose workflow in test/cluster/mzcompose.py. The test workflow exercises environmentd crashes with and without the existence of user-defined objects in the schema.