compute-client: refactor controller errors
This commit refactors the error types returned by the compute controller
methods. Instead of returning a big-ball-of-mud `ComputeError` from
every method, making it impossible for callers to see which of the many
error variants each method can actually produce, we now return bespoke
error types that precisely enumerate each method's possible errors.

The change introduces some boilerplate code because Rust does not give
us efficient means to deal with enums that have intersecting sets of
variants. The `thiserror` crate makes the error definitions a bit less
verbose. On the plus side, the compute controller API is now easier to
use correctly. And being specific about error types also forces us to
think about the errors we actually want to return to callers. Prior to
this change, it was easy to just slap a `?` at the end of the line and
consider the error properly handled, even though it might have been the
right thing to panic or ignore the error instead.