Monitoring of `computed` compute command history size (#16009)
This commit introduces compute metrics to aid in monitoring the computed compute command history size as well as utilizes these metrics in a regression test to ensure that the compute command history size remains well-behaved when hundreds of fast- and slow-path peeks are executed. The compute metrics are exposed through Prometheus and comprise the raw length of the compute command history as well as the number of dataflows presently in the history. The former metric varies according to the garbage collection policy of reducing the history upon doublings; the latter is more stable and tracks the logical clean-up of dataflows in the history (e.g., created as a result of slow-path peeks) whenever the history is reduced.