sql[-parser]: remove STATISTICS INTERVAL MS Kafka option
The STATISTICS INTERVAL MS option controls how often librdkafka
statistics are dumped to stderr via an `info!` log. This is no longer a
useful option to expose to end users, as they do not have access to
Materialize's stderr logs. So just remove it.

In the future, we may want to force-enable the statistics callback with
something smarter, that extracts relevant stats in librdkafka as
Prometheus metrics in Materialize, but that's a project for another day.

Fix #16050.