deps: bump to latest tokio and tokio-util
Due to a change in Tokio v1.21, this upgrade required a tweak to how we
handle checking for closed TCP streams during `COPY ... TO` operations.
I didn't bother digging into cause of the change, but the observed
behavior change is that `conn.ready(Interest::READABLE)` is no longer
guaranteed to return quickly, even if the TCP stream is healthy.

The fix is to move the call to `conn.ready` to the left-hand side of the
select!--i.e., to the select! condition rather than the select handler.
That way a stalled call to `conn.ready` on a healthy TCP stream will not
stall the connection handling as a whole, but a connection close event
will still unstall the call to `conn.ready` and cause the connection to
be terminated and cleaned up.

As an added bonus, the fix unifies the handling of closed connections
across `COPY ... TO` and normal `SELECT`s.

This commit also updates our usage of unstable Tokio APIs to restore
compatibility with the Tokio console (i.e., building with the
`--tokio-console` flag).

Co-authored-by: Jan Teske <jteske@posteo.net>