test: fix select_introspection.slt
The SLT testing replica-targeted peeks broke recently because of changes
we made to how replica-targeted peeks work (previously, only one replica
would handle the peek, now they all handle it but only one response is
forwarded). This commit fixes that test and cleans up the SLT file in
the process.

Querying introspection values in SLTs is problematic because it relies
on timing. If the introspection results are not immediately available,
the test becomes flaky. To solve this, this commit moves the check that
has to look at the introspection results into testdrive instead.