storage: deprovision storage hosts in `process`
This commit moves the deprovisioning of storage hosts out of
`update_read_capabilities` (and `drop_sinks_unvalidated`) into the
`process` method. The main benefit of this is that
`update_read_capabilities` becomes sync, which has a significant effect
on the compute controller API.

This commit only makes the changes to host deprovisioning. Subsequent
commits propagate their consequences to the rest of the code base.