coord: Initialize new tables uppers (#16160)
When creating a new collection, the upper is initialized to the
minimum timestamp value. The collection is not readable until the upper
is advanced.

Previously, when creating a table we don't immediately advance it's
upper. As a consequence the table doesn't become readable until the
next periodic table advancement.

This commit updates the create table DDL logic, so that the table's
upper is immediately advanced, and therefore immediately readable.

Fixes #16115