persist: admin tool to manually compact a shard
Persist shards currently only perform compactions that they directly
caused by inserting a batch. Any others are assumed to be done by
someone else. Among other things, this doesn't backfill compactions when
we change our selection criteria (such as when we started considering
batch parts). This is one of the things we'll be revisiting in Persist
Compaction 2.0, but until then add an admin tool we can use to manually
fixup individual shards.

Touches #16161