test: Panic in test instead of returning an error (#16248)
This commit modifies all the integration tests in mz-environmentd, so
instead of returning an error they panic. When a test returns an error
you lose the stack trace of where that error was returned from, which
makes it extremely difficult to debug certain test failures. This is
especially true of flaky tests in CI. Panics will print the full stack
trace of where the test panicked from.