adapter: print COUNT(*) when EXPLAIN'ing COUNT(*)
Early on, we turn COUNT(*) into COUNT(true), which is then a bit
surprising in EXPLAIN output.  It seems less surprising to print
COUNT(*) for this case, since that's far more likely to be what the
user actually typed than COUNT(true).

Note that Postgres treats COUNT(*) as COUNT with zero arguments (see
aggr_arg in the grammar and its associated comment), which might make
distinguishing this case cleaner than what we do right now.

It would be nice to avoid duplicating is_count_asterisk(), but my
attempts to de-duplicate it ended up much less readable with all the
necessary manipulations to extract the common bits.

Fixes #15939.