storage: correctly calculate the times to reclock to
The logic in `ReclockOperator::reclock_time`, even though it was
thoroughly documented and passed numerous tests, was not correct and
could end up reclocking `FromTime` times earlier than they should. Its
fatal flaw was that it was assuming that one of these two cases must
always hold:
  * When `src_ts` changes from being beyond the `FromTime` frontier to
    not beyond we will observe a triplet with a positive diff whose
    `from_ts` is not beyond `src_ts`.
  * When the frontier represented by the remap collection advances but
    it `src_ts` remains beyond it we will observe a triplet with a
    positive diff whose `from_ts` is beyond `src_ts`.

Both of the above assumptions turn out to be incorrect.

The first assumption is not true if the `FromTime` frontier ever
advances to the empty antichain. In that case there is no positive diff
to observe since there are no timestamps left in the frontier. We have
not observed this happening because all production sources never reach
the empty antichain.

The second assumption is not true if the `FromTime` frontier advances
but in an unrelated to `src_ts` way. For example, if `src_ts` is about
partition 0 of a Kafka topic it's possible that we see a lot of remap
updates for a different partition. The current code would incorrectly
reclock `src_ts` to one of these irrelevant frontier advancements
because it would assume that the whole frontier moved past `src_ts`
since `src_ts` was not beyond all of the updates that happened. This was
the cause of #16318.

In retrospect, the implementation that filtered diffs and used
`i64::MIN` was very sucpicious but the extensive unit tests of the
reclock logic and a green CI convinced me of its correctness.

The new logic implemented in this PR seems much more "natural" without
relying on weird minimum values or filtering diffs. It works correctly
on both of the edge cases illustrated above and passed all previous
tests without modification.

Fixes #16318

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>