persist_sink: state-of-the-art async operator
This commit migrates the compute persist_sink operators to the "new way"
of writing async timely operators, namely using `AsyncOperatorBuilder`
rather than the `build_async` extension method.

The immediate benefit that gives us is that we can make use of the shutdown
button returned by `AsyncOperatorBuilder::build` now, to reliably clean up
the dataflow once its token is dropped. Prior to this change, operators
would only shut down if they were scheduled by timely again after the
token was dropped, because the token check happened inside the operator
logic.

As a result of the above, dropping a materialized view that has advanced
to the empty frontier now cleans up the persist_sink operators, so they are
not leaked anymore. Ideally, we'd want them to be cleaned up as soon as
their inputs advanced to the empty frontier. This will be taken care of
by a later commit.

This commit also leaves heartbeating the persist client temporarily
broken. We will also deal with this in a later commit.