timely-util: shutdown button for all async operators
This PR makes the AsyncOperatorBuilder to always return a
`ShutdownButton` as part of its construction. This button can either be
ignored or converted into a deadman's switch which activates on `Drop`
by calling `ShutdownButton::press_on_drop`.

This allows anyone who needs to create a tokened operator to use this
button and end up with a cleaner structure of the async logic.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>