adapter: Fix builtin schema migrations (#16402)
Previously, the adapter was only updating the persisted definition of
system objects that were directly migrated. If a system object was
migrated because it depended on an upstream system object that was
migrated, then its persisted definition was not updated. In these
scenarios, the migrated system object definition would be lost after
a system restart, leading to an inconsistent stash.

This commit updates the builtin migration framework so that the
persisted definitions of all system objects are updated.

Fixes #16396