[sinks] Limit the number of retries on retriable transaction errors (#16374)
* Limit the number of retries on retriable transaction errors

The design philosophy for sinks is that, on persistent errors, the sink
should eventually halt. This is especially important now that we track
the status of the sink, so errors are reported promptly.

Anyways, now we have a smallish retry limit on each of the two
operations. I've also converted these back to a more functional style,
since otherwise it'd be tedious to track the previous error across loop
iterations.

* ore: teach retry module about fatal errors

Not all errors are retryable. This commit teaches mz_ore::retry::Retry
about trivariate retry statuses: Ok, RetryableErr, and FatalErr. Both Ok
and FatalErr immediately end the retry operation, and return Ok or Err
to the caller, respectively, while onlya RetryableErr is retried
according to the retry operation configuration.

* Don't bother retrying the transaction abort

And also leave a comment, since it's not obvious why it's worth the
trouble to do at all.

Co-authored-by: Nikhil Benesch <nikhil.benesch@gmail.com>