adapter: Fix storage collection usage timestamps (#16429)
Previously, the timestamp used for storage collection usage was taken
after the usage was retrieved from persist. This caused the timestamps
to slowly (sometimes quickly) drift apart because it didn't account for
the time take to collect usage metrics.

This commit moves the timestamp so that it's taken before the storage
usage is retrieved. This helps ensure that there's an equal interval
between all storage usage metrics.