stash: verify and seal in known order
Previously we would verify the old upper and seal (advance the upper) in
separate futures. If sealing executed before the check, the check would
fail. We've long had a policy to assume that futures could first-poll or
execute in any order. This is the first suspected example of incorrectly
violating that policy resulting in production outages.

Fixes #16425