compute-client: move public command types into separate module
This commit moves `DataflowDescription` and its contained types out of
`mz_compute_client::command` into a new module
`mz_compute_client::types` (inspired by the same module in
`mz_storage_client`).

There are two motivations for this change:
 * Reducing the size of the `command` module.
 * Avoiding the need for clients to import from the `command` module.
   This module defines (part of) the compute protocol and should only be
   imported by `mz-compute-client` and `mz-compute`.

The second goal is not quite reached with this refactor, as some ID
types (`ReplicaId` and `ProcessId`) are also defined in the `command`
module. This will be resolved in another commit.