Unrevert "persist: read a bounded # of rows from CRDB when fetching latest state (#16143)"
This reverts commit 6663bde24a5696680b34a0540d5f47c114c52575, which was
a revert of #16143. It also fixes the issue that led to the revert,
which was a race condition in the order of blob and consensus wipes
involved in `./bin/environmentd --reset`. Concretely, if we wipe blob
(files) first and then consensus (crdb/postgres) and are interrupted in
between, then we could end up with consensus state pointing at no longer
existing files. There's no race if we wipe them in the opposite order.