*: disallow potentially dangerous `as` conversions
The `as` operator silently truncates data in many situations. It is very
difficult to assess whether a given usage of `as` is dangerous or not.
This commit bans its usage across the codebase, to force usage of safer
patterns, like `From` and `TryFrom`.

There are a few modules, functions, and statements that are difficult or
impossible to rewrite to avoid `as`. These sites have been opted out of
the ban via `#[allow(clippy::as_conversion)]`.