Adapter: Fix handling of constant queries in txns (#16488)
Constant queries are queries that do not involve any data that
originates from a source. For example these involve queries like
`SELECT 1`, `SELECT mz_now()`, or `SELECT * FROM v` where v is a view
that doesn't contain any sources.

These queries can further be broken down into two categories.
  1) Timestamp/Timeline Dependent: The result of these queries depend
  on the timeline that they are executed in and the timestamp selected.

  2) Timestamp/Timeline Independent: The result of these queries do not
  depend on the timeline that they are executed in nor the timestamp
  selected.

Both types of constant queries should be able to be run in any
transaction, regardless of the transaction's timeline/timedomain.

Previously, we had two bugs with these types of queries.
  1) Timestamp/Timeline Dependent queries would cause panics in
  transactions that had already selected a timeline/timedomain.

  2) Views and Materialized Views that were Timestamp/Timeline
  Independent were mistakenly classified as Timestamp/Timeline
  dependent.

This commit fixes both these issues. Additionally, it refactors much
of the timestamp and timeline related code to make the different types
of queries more explicit.

Additionally, it makes the change that if the timeline of a constant
Timestamp/Timeline dependent view is not specified, then it defaults
to the epoch millisecond timeline.

Fixes #16488