[docs] Document more of our intended Kafka semantics (#16483)
* Document more of our intended Kafka semantics

We'd like to say something useful about our sink semantics. (If we
didn't think it was important to avoid duplicates, the sink code would
be much simpler!) However, we don't want to commit to more than we can
promise, and we don't have much control over either Kafka itself or the
specific cluster that MZ is connected to.

This tries to balance the concerns by using Kafka's language for our
guarantees (vs. trying to capture Kafka's guarantees in database
terminology) and emphasizing the end-to-end nature of an exactly-once
processing guarantee and which parts are out of MZ's control.

* Move content around after restructure

Co-authored-by: morsapaes <marta.paes.moreira@gmail.com>