ci: re-enable cargo test in CI
After changing to nexttext in 0275079d429fa669bb497b78405840eb59ac910a
the CI no longer ran the `cargo test` job because it was using the
default `glob` to detect if it should run it, which never matched any
files.

The pipeline previously defined the `cargo test` step as a dependency of
the `build-x86` step, which had a correct input definition and therefore
was always triggered when the other job triggered.

This PR sets the `inputs` of the `cargo test` job to be the same as the
build job so that the previous behaviour is maintained (tests run when
the build runs.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>