persist: disconnect the flow_control_input from shard_source_descs
We could leave it connected without risking deadlock so long as there is
a non zero summary on the feedback edge. But it may be less efficient
because the pipeline will be moving in increments of SUMMARY even though
we have potentially dumped a lot more data in the pipeline because of
batch boundaries. Leaving it unconnected means the pipeline will be able
to retire bigger chunks of work.