mzcompose: clear stale process metadata via tmpfs
The process orchestrator stores per-incarnation state in /tmp (e.g.,
process IDs). After a restart, the orchestrator expects that /tmp has
been cleared by the OS, as the metadata is no longer valid. This is not
true with Docker containers, however; the /tmp directory survives a
container stop and start.

We were previously working around this by manually deleting the metadata
on a restart, with `rm -rf`. Per #16746, this approach was blowing away
some important log files. Change to a different approach, which installs
a temporary filesystem ("tmpfs") on /tmp. This temporary filesystem does
*not* survive a container restart, as desired.

Fix #16746.