persist: don't panic in maybe_heartbeat_* for tombstone shards
It's probably surprising to the caller that the shard becoming a
tombstone expired the reader/writer. Possibly the right thing to do here
is pass up a bool to the caller indicating whether the id it's trying to
heartbeat has been expired, but that happening on a tombstone vs not is
very different. As a medium-term compromise, pretend we did the
heartbeat here.