timely-util: define `ProbeNotify` trait
This PR defices a `ProbeNotify` trait that is available for all timely
Streams which is functionally identical to the built-in `Probe` trait
with the exception that it supports progress notifications to consumers
of the handle.

This is done via the `Handle::progressed()` async function which will
resolve when the managed frontier has made progress.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>