add an Active Replication reference document (#16777)
This commit proposes a first version of a reference document describing our active replication mechanism. The document comprises a high-level and brief description of the mechanism as well as links to the relevant code. The extended treatment of several related concepts (e.g., read and write frontier management for compute collections, history equivalence and reduction, and reconciliation) is linked to in other documents, which are at this point only filled out with TODO annotations for future work.

So in short, the changes in this commit are:

* A first version of an active replication document for the reference documentation of the COMPUTE layer.
* Linking to this document from the README.md of the COMPUTE layer.
* A number of other documents for future work on: (a) read and write frontier management for compute collections, (b) reconciliation, (c) compute command model.