persist: fix bug causing accidentally serial compaction
We were accidentally serializing all compaction work in the
PersistCompactionWorker task. The intent was to limit concurrency of the
actual work using a semaphore, enqueue requests in a channel leading
into that semaphore, and drop requests when the channel fills up (i.e.
both the semaphore and the channel are full). This commit changes the
impl to match that intent.

Touches #15093