orchestrator-process: allocate TCP->UDS proxies for services
We switched the process orchestrator to use Unix domain sockets rather
than TCP in ac1028d87a2, to avoid race conditions when allocating TCP
ports. The conversion has vastly improved the reliability of the process
orchestrator itself, but has the rather large downside that many other
tools (e.g., Prometheus, web browsers) are not compatible with Unix
domain sockets.

This commit teaches the process orchestrator a new trick: the ability to
allocate a TCP proxy for each UDS socket. For each named port for each
created service, the process orchestrator allocates a free TCP port,
prints the allocated port to the log file, then forwards traffic on that
port to the UDS socket.

This gives us the best of all worlds. Intra-Materialize communication
occurs over UDS, with no port TCP port allocation issues, but there is a
(randomly assigned) TCP port available for live debugging of a service
via Prometheus or a web browser.

h/t @danhhz for helping to design this approach.